+++ files/lib/page/ConversationListPage.class.php
@@ -7,12 +7,13 @@ use wcf\system\clipboard\ClipboardHandler;
 use wcf\system\exception\IllegalLinkException;
 use wcf\system\request\LinkHandler;
 use wcf\system\WCF;
+use wcf\util\ArrayUtil;
 
 /**
  * Shows a list of conversations.
  * 
  * @author	Marcel Werk
- * @copyright	2001-2012 WoltLab GmbH
+ * @copyright	2001-2016 WoltLab GmbH
  * @license	GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @package	com.woltlab.wcf.conversation
  * @subpackage	page
@@ -20,44 +21,44 @@ use wcf\system\WCF;
  */
 class ConversationListPage extends SortablePage {
 	/**
-	 * @see	\wcf\page\AbstractPage::$enableTracking
+	 * @inheritDoc
 	 */
 	public $enableTracking = true;
 	
 	/**
-	 * @see	\wcf\page\SortablePage::$defaultSortField
+	 * @inheritDoc
 	 */
 	public $defaultSortField = CONVERSATION_LIST_DEFAULT_SORT_FIELD;
 	
 	/**
-	 * @see	\wcf\page\SortablePage::$defaultSortOrder
+	 * @inheritDoc
 	 */
 	public $defaultSortOrder = CONVERSATION_LIST_DEFAULT_SORT_ORDER;
 	
 	/**
-	 * @see	\wcf\page\SortablePage::$validSortFields
+	 * @inheritDoc
 	 */
-	public $validSortFields = array('subject', 'time', 'username', 'lastPostTime', 'replies', 'participants');
+	public $validSortFields = ['subject', 'time', 'username', 'lastPostTime', 'replies', 'participants'];
 	
 	/**
-	 * @see	\wcf\page\MultipleLinkPage::$itemsPerPage
+	 * @inheritDoc
 	 */
-	public $itemsPerPage = CONVERSATIONS_PER_PAGE;
+	public $itemsPerPage = 1;
 	
 	/**
-	 * @see	\wcf\page\AbstractPage::$loginRequired
+	 * @inheritDoc
 	 */
 	public $loginRequired = true;
 	
 	/**
-	 * @see	\wcf\page\AbstractPage::$neededModules
+	 * @inheritDoc
 	 */
-	public $neededModules = array('MODULE_CONVERSATION');
+	public $neededModules = ['MODULE_CONVERSATION'];
 	
 	/**
-	 * @see	\wcf\page\AbstractPage::$neededPermissions
+	 * @inheritDoc
 	 */
-	public $neededPermissions = array('user.conversation.canUseConversation');
+	public $neededPermissions = ['user.conversation.canUseConversation'];
 	
 	/**
 	 * list filter
@@ -73,7 +74,7 @@ class ConversationListPage extends SortablePage {
 	
 	/**
 	 * label list object
-	 * @var	\wcf\data\conversation\label\ConversationLabelList
+	 * @var	ConversationLabelList
 	 */
 	public $labelList = null;
 	
@@ -102,7 +103,13 @@ class ConversationListPage extends SortablePage {
 	public $outboxCount = 0;
 	
 	/**
-	 * @see	\wcf\page\IPage::readParameters()
+	 * participant that
+	 * @var	string[]
+	 */
+	public $participants = [];
+	
+	/**
+	 * @inheritDoc
 	 */
 	public function readParameters() {
 		parent::readParameters();
@@ -130,18 +137,27 @@ class ConversationListPage extends SortablePage {
 				throw new IllegalLinkException();
 			}
 		}
+		
+		if (isset($_REQUEST['participants'])) $this->participants = ArrayUtil::trim(explode(',', $_REQUEST['participants']));
 	}
 	
 	/**
-	 * @see	\wcf\page\MultipleLinkPage::initObjectList()
+	 * @inheritDoc
 	 */
 	protected function initObjectList() {
 		$this->objectList = new UserConversationList(WCF::getUser()->userID, $this->filter, $this->labelID);
 		$this->objectList->setLabelList($this->labelList);
+		
+		if (!empty($this->participants)) {
+			$this->objectList->getConditionBuilder()->add('conversation.conversationID IN (SELECT conversationID FROM wcf'.WCF_N.'_conversation_to_user WHERE username IN (?) GROUP BY conversationID HAVING COUNT(conversationID) = ?)', [
+				$this->participants,
+				count($this->participants)
+			]);
+		}
 	}
 	
 	/**
-	 * @see	\wcf\page\IPage::readData()
+	 * @inheritDoc
 	 */
 	public function readData() {
 		parent::readData();
@@ -152,7 +168,7 @@ class ConversationListPage extends SortablePage {
 		}
 		
 		// read stats
-		if (!$this->labelID) {
+		if (!$this->labelID && empty($this->participants)) {
 			switch ($this->filter) {
 				case '':
 					$this->conversationCount = $this->items;
@@ -172,31 +188,31 @@ class ConversationListPage extends SortablePage {
 			}
 		}
 		
-		if ($this->filter != '' || $this->labelID) {
+		if ($this->filter != '' || $this->labelID || !empty($this->participants)) {
 			$conversationList = new UserConversationList(WCF::getUser()->userID, '');
 			$this->conversationCount = $conversationList->countObjects();
 		}
-		if ($this->filter != 'draft' || $this->labelID) {
+		if ($this->filter != 'draft' || $this->labelID || !empty($this->participants)) {
 			$conversationList = new UserConversationList(WCF::getUser()->userID, 'draft');
 			$this->draftCount = $conversationList->countObjects();
 		}
-		if ($this->filter != 'hidden' || $this->labelID) {
+		if ($this->filter != 'hidden' || $this->labelID || !empty($this->participants)) {
 			$conversationList = new UserConversationList(WCF::getUser()->userID, 'hidden');
 			$this->hiddenCount = $conversationList->countObjects();
 		}
-		if ($this->filter != 'outbox' || $this->labelID) {
+		if ($this->filter != 'outbox' || $this->labelID || !empty($this->participants)) {
 			$conversationList = new UserConversationList(WCF::getUser()->userID, 'outbox');
 			$this->outboxCount = $conversationList->countObjects();
 		}
 	}
 	
 	/**
-	 * @see	\wcf\page\IPage::assignVariables()
+	 * @inheritDoc
 	 */
 	public function assignVariables() {
 		parent::assignVariables();
 		
-		WCF::getTPL()->assign(array(
+		WCF::getTPL()->assign([
 			'filter' => $this->filter,
 			'hasMarkedItems' => ClipboardHandler::getInstance()->hasMarkedItems(ClipboardHandler::getInstance()->getObjectTypeID('com.woltlab.wcf.conversation.conversation')),
 			'labelID' => $this->labelID,
@@ -204,7 +220,8 @@ class ConversationListPage extends SortablePage {
 			'conversationCount' => $this->conversationCount,
 			'draftCount' => $this->draftCount,
 			'hiddenCount' => $this->hiddenCount,
-			'outboxCount' => $this->outboxCount
-		));
+			'outboxCount' => $this->outboxCount,
+			'participants' => $this->participants
+		]);
 	}
 }