+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### 2.2.0 Alpha 1 (XXXX-YY-ZZ)
 
 * Only search active conversation when using search bar in a conversation.
+* Filter option for conversation participants on conversation list page.